+++ lib/toc_doc/models.rb
@@ -1,10 +1,11 @@
 # frozen_string_literal: true
 
 require 'toc_doc/models/resource'
+require 'toc_doc/models/speciality'
+require 'toc_doc/models/place'
 require 'toc_doc/models/search'
 
 require 'toc_doc/models/availability'
 require 'toc_doc/models/availability/collection'
 
 require 'toc_doc/models/profile'
-require 'toc_doc/models/speciality'